refactor: useAddressName hook
